requirements.txt
streamlit>=1.30.0
chromadb>=0.4.22
groq>=0.4.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.3.0
